tests/wpt/web-platform-tests/svg/painting/parsing/stroke-opacity-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: getComputedStyle().strokeOpacity</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeOpacityProperty"/>
    <h:meta name="assert" content="stroke-opacity computed value is clamped to the range [0,1]."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("stroke-opacity", "-1", "0");
test_computed_value("stroke-opacity", "0.5");
test_computed_value("stroke-opacity", "3", "1");
test_computed_value("stroke-opacity", "-100%", "0");
test_computed_value("stroke-opacity", "50%", "0.5");
test_computed_value("stroke-opacity", "300%", "1");

  ]]></script>
</svg>
